Time: 491 ms (57.14%), Space: 300.7 MB (6.67%) - LeetHub

+++ 3646-sum-of-good-subsequences/3646-sum-of-good-subsequences.cpp
@@ -15,7 +15,7 @@ public:
             {
                 count[nums[i]] += (1 + count[nums[i]+1]%MOD + count[nums[i]-1]%MOD)%MOD;
                 count[nums[i]] %= MOD;
-                sum[nums[i]] +=(int)(nums[i] + (sum[nums[i]+1]%MOD + count[nums[i]+1]*1LL*nums[i]%MOD)%MOD+(sum[nums[i]-1]%MOD + count[nums[i]-1]*1LL*nums[i]%MOD)%MOD)%MOD;
+                sum[nums[i]] +=(nums[i] + (sum[nums[i]+1]%MOD + count[nums[i]+1]*1LL*nums[i]%MOD)%MOD+(sum[nums[i]-1]%MOD + count[nums[i]-1]*1LL*nums[i]%MOD)%MOD)%MOD;
                 sum[nums[i]] %= MOD;
 
             }
